Add KubernetesPluginConfig and move deploy target config
