+++ pkg/app/pipedv1/plugin/kubernetes/config/application.go
@@ -97,39 +97,6 @@ type KubernetesVariantLabel struct {
 	BaselineValue string `json:"baselineValue" default:"baseline"`
 }
 
-type KubernetesDeployTargetConfig struct {
-	// The master URL of the kubernetes cluster.
-	// Empty means in-cluster.
-	MasterURL string `json:"masterURL,omitempty"`
-	// The path to the kubeconfig file.
-	// Empty means in-cluster.
-	KubeConfigPath string `json:"kubeConfigPath,omitempty"`
-	// Version of kubectl will be used.
-	KubectlVersion string `json:"kubectlVersion"`
-	// Configuration for application resource informer.
-	AppStateInformer KubernetesAppStateInformer `json:"appStateInformer"`
-}
-
-// KubernetesAppStateInformer represents the configuration for application resource informer.
-type KubernetesAppStateInformer struct {
-	// Only watches the specified namespace.
-	// Empty means watching all namespaces.
-	Namespace string `json:"namespace,omitempty"`
-	// List of resources that should be added to the watching targets.
-	IncludeResources []KubernetesResourceMatcher `json:"includeResources,omitempty"`
-	// List of resources that should be ignored from the watching targets.
-	ExcludeResources []KubernetesResourceMatcher `json:"excludeResources,omitempty"`
-}
-
-// KubernetesResourceMatcher represents the matcher for a Kubernetes resource.
-type KubernetesResourceMatcher struct {
-	// The APIVersion of the kubernetes resource.
-	APIVersion string `json:"apiVersion,omitempty"`
-	// The kind name of the kubernetes resource.
-	// Empty means all kinds are matching.
-	Kind string `json:"kind,omitempty"`
-}
-
 // K8sResourcePatch represents a patch operation for a Kubernetes resource.
 type K8sResourcePatch struct {
 	// The target of the patch operation.
